+++ HISTORY.rst
@@ -2,9 +2,12 @@
 History
 =======
 
+.. py:module:: zammadoo
+
 future release
 --------------
-* **[breaking change]** Client parameter: instead ``username`` and ``password`` now ``http_auth=(username, password)`` is provided
+* **breaking change**: :class:`client.Client` now uses ``http_auth=(username, password)`` as authentication parameter
+* **new feature**: :meth:`tickets.Ticket.history` method
 
 0.1.0 (2023-10-08)
 ------------------